LUCENE-10436: Reinstate public getdocValuesdocIdSetIterator method on DocValues
